Missing S3 Object Should Raise an Exception
When the Object containing the SSA response, as designated in the
SQS message, is missing from the S3 bucket, we should not simply
issue a warning message.  This should be considered an Exception -
waiting for the Object to show up is not going to do anyone any good.
This was encountered during testing by using an incorrect bucket name,
but if the bucket name were correct we would continue looking for the
object until the request timed out.